tests: move functional tests in `test_edit_map.py`.
